+++ apps/client/src/app/pages/allagan-reports/allagan-report-row/allagan-report-row.component.ts
@@ -56,6 +56,9 @@ export class AllaganReportRowComponent {
   delete = new EventEmitter<void>();
 
   @Output()
+  deleteOwn = new EventEmitter<void>();
+
+  @Output()
   edit = new EventEmitter<void>();
 
   applyingChange = false;