+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `f71601effc96842393f04b71bf33f54f65c04b5e`;
+export const extractsHash = `0ee9232afeb6c86174224a00ccb32688f1dd81fa`;