+++ apps/data-extraction/src/extractors/recipes.extractor.ts
@@ -11,7 +11,7 @@ export class RecipesExtractor extends AbstractExtractor {
     };
     combineLatest([
       this.getAllEntries('https://xivapi.com/CompanyCraftSequence'),
-      this.aggregateAllPages('https://xivapi.com/Recipe?columns=ID,ClassJob.ID,MaterialQualityFactor,DurabilityFactor,QualityFactor,DifficultyFactor,RequiredControl,RequiredCraftsmanship,CanQuickSynth,RecipeLevelTable,AmountResult,ItemResultTargetID,ItemIngredient0,ItemIngredient1,ItemIngredient2,ItemIngredient3,ItemIngredient4,ItemIngredient5,ItemIngredient6,ItemIngredient7,ItemIngredient8,ItemIngredient9,AmountIngredient0,AmountIngredient1,AmountIngredient2,AmountIngredient3,AmountIngredient4,AmountIngredient5,AmountIngredient6,AmountIngredient7,AmountIngredient8,AmountIngredient9,IsExpert,SecretRecipeBookTargetID')
+      this.aggregateAllPages('https://xivapi.com/Recipe?columns=ID,ClassJob.ID,MaterialQualityFactor,DurabilityFactor,QualityFactor,DifficultyFactor,RequiredControl,RequiredCraftsmanship,CanQuickSynth,RecipeLevelTable,AmountResult,ItemResultTargetID,ItemIngredient0,ItemIngredient1,ItemIngredient2,ItemIngredient3,ItemIngredient4,ItemIngredient5,ItemIngredient6,ItemIngredient7,ItemIngredient8,ItemIngredient9,AmountIngredient0,AmountIngredient1,AmountIngredient2,AmountIngredient3,AmountIngredient4,AmountIngredient5,AmountIngredient6,AmountIngredient7,AmountIngredient8,AmountIngredient9,IsExpert,SecretRecipeBook')
     ]).subscribe(([companyCrafts, xivapiRecipes]) => {
       xivapiRecipes.forEach(recipe => {
         if (recipe.RecipeLevelTable === null) {
@@ -49,7 +49,7 @@ export class RecipesExtractor extends AbstractExtractor {
           controlReq: recipe.RequiredControl,
           craftsmanshipReq: recipe.RequiredCraftsmanship,
           rlvl: recipe.RecipeLevelTable.ID,
-          masterbook: recipe.SecretRecipeBookTargetID,
+          masterbook: recipe.SecretRecipeBook?.ItemTargetID,
           ingredients: ingredients
             .map(ingredient => {
               return {